notificationScript/requirements.txt
python-dotenv
requests
beautifulsoup4
plyer
flask
